requirements.txt
discord.py==2.7.1
openai>=1.54.0
python-dotenv==1.0.0
duckduckgo-search
